requirements.txt
# Core ML Frameworks
torch>=2.1.0
transformers>=4.36.0
peft>=0.7.0
trl>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0

# Data Processing
datasets>=2.15.0
pandas>=2.0.0
numpy>=1.24.0

# Configuration Management
pydantic>=2.0.0
pyyaml>=6.0

# Utilities
scikit-learn>=1.3.0
tqdm>=4.65.0

# JSON Repair for resilient parsing
json-repair>=0.7.0

# Logging
wandb>=0.15.0  # Optional: for experiment tracking
tensorboard>=2.14.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
